// src/app/models/organisation.ts
export class Membre {
    profile: string;
    experiences: [];
    educations: [];
    nationalite: string;
    localisation: string;
    phone : string;
    poste : string; 
    resume : string;
    facebook: string;
    twitter : string;
    instagram : string; 
    linkedin : string;
    constructor( public name: string, public age :string, public email : string){}
}